pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "infra-ai"
version = "0.1.0"
description = "LangGraph workflow from requirements to validated infra configs and generated code"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "langgraph>=0.2.60",
    "langchain-core>=0.3.30",
    "langchain-openai>=0.3.0",
    "langchain-ollama>=0.2.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "pyyaml>=6.0.2",
    "gitpython>=3.1.43",
    "httpx>=0.28.0",
]

[project.optional-dependencies]
milvus = ["pymilvus>=2.4.0", "sentence-transformers>=3.0.0"]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
]

[project.scripts]
infra-ai = "infra_ai.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["infra_ai"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"
